+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Platform/Vault/New-AzDataProtectionBackupVault.ps1
@@ -49,6 +49,16 @@ function New-AzDataProtectionBackupVault
         [ValidateSet('Disabled','Unlocked', 'Locked')]
         ${ImmutabilityState},
 
+        [Parameter(Mandatory=$false, HelpMessage='Immutability type of the vault. Allowed values are AsPerPolicy and TimeBased.')]
+        [System.String]
+        [ValidateSet('AsPerPolicy', 'TimeBased')]
+        ${ImmutabilityType},
+
+        [Parameter(Mandatory=$false, HelpMessage='Immutability duration in days. Required when ImmutabilityType is TimeBased.')]
+        [System.Double]
+        [ValidateRange(30, 36135)]
+        ${ImmutabilityDurationInDay},
+
         [Parameter(Mandatory=$false, HelpMessage='Cost Management Granularity of the vault. Allowed values are VaultLevel, ProtectedItemLevel, ProtectedItemWithParentTag.')]
         [System.String]
         [ValidateSet('VaultLevel', 'ProtectedItemLevel', 'ProtectedItemWithParentTag')]
@@ -154,6 +164,31 @@ function New-AzDataProtectionBackupVault
 
     process
     {
+        $hasImmutabilityState = $PSBoundParameters.ContainsKey("ImmutabilityState")
+        $hasImmutabilityType = $PSBoundParameters.Remove("ImmutabilityType")
+        $hasImmutabilityDurationInDay = $PSBoundParameters.Remove("ImmutabilityDurationInDay")
+
+        Assert-AzDataProtectionImmutabilitySetting `
+            -ImmutabilityState $ImmutabilityState `
+            -ImmutabilityType $ImmutabilityType `
+            -HasImmutabilityState $hasImmutabilityState `
+            -HasImmutabilityType $hasImmutabilityType `
+            -HasImmutabilityDurationInDay $hasImmutabilityDurationInDay
+
+        if ($hasImmutabilityType -or $hasImmutabilityDurationInDay) {
+            $immutabilityPipeline = Get-AzDataProtectionImmutabilityRequestPipeline `
+                -ImmutabilityType $ImmutabilityType `
+                -ImmutabilityDurationInDay $ImmutabilityDurationInDay `
+                -HasImmutabilityType $hasImmutabilityType `
+                -HasImmutabilityDurationInDay $hasImmutabilityDurationInDay
+
+            if ($PSBoundParameters.ContainsKey("HttpPipelinePrepend")) {
+                $PSBoundParameters["HttpPipelinePrepend"] = $PSBoundParameters["HttpPipelinePrepend"] + @($immutabilityPipeline)
+            }
+            else {
+                $PSBoundParameters.Add("HttpPipelinePrepend", @($immutabilityPipeline))
+            }
+        }
 
         $hasCmkEncryptionState = $PSBoundParameters.Remove("CmkEncryptionState")
         $hasCmkIdentityType = $PSBoundParameters.Remove("CmkIdentityType")