+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Platform/Vault/Update-AzDataProtectionBackupVault.ps1
@@ -36,6 +36,16 @@ function Update-AzDataProtectionBackupVault
         [ValidateSet('Disabled','Unlocked', 'Locked')]
         ${ImmutabilityState},
 
+        [Parameter(ParameterSetName="UpdateExpanded",Mandatory=$false, HelpMessage='Immutability type of the vault. Allowed values are AsPerPolicy and TimeBased.')]
+        [System.String]
+        [ValidateSet('AsPerPolicy', 'TimeBased')]
+        ${ImmutabilityType},
+
+        [Parameter(ParameterSetName="UpdateExpanded",Mandatory=$false, HelpMessage='Immutability duration in days. Required when ImmutabilityType is TimeBased.')]
+        [System.Double]
+        [ValidateRange(30, 36135)]
+        ${ImmutabilityDurationInDay},
+
         [Parameter(ParameterSetName="UpdateExpanded",Mandatory=$false, HelpMessage='Cost Management Granularity of the vault. Allowed values are VaultLevel, ProtectedItemLevel, ProtectedItemWithParentTag.')]
         [System.String]
         [ValidateSet('VaultLevel', 'ProtectedItemLevel', 'ProtectedItemWithParentTag')]
@@ -150,6 +160,32 @@ function Update-AzDataProtectionBackupVault
 
     process
     {
+        $hasImmutabilityState = $PSBoundParameters.ContainsKey("ImmutabilityState")
+        $hasImmutabilityType = $PSBoundParameters.Remove("ImmutabilityType")
+        $hasImmutabilityDurationInDay = $PSBoundParameters.Remove("ImmutabilityDurationInDay")
+
+        Assert-AzDataProtectionImmutabilitySetting `
+            -ImmutabilityState $ImmutabilityState `
+            -ImmutabilityType $ImmutabilityType `
+            -HasImmutabilityState $hasImmutabilityState `
+            -HasImmutabilityType $hasImmutabilityType `
+            -HasImmutabilityDurationInDay $hasImmutabilityDurationInDay
+
+        if ($hasImmutabilityType -or $hasImmutabilityDurationInDay) {
+            $immutabilityPipeline = Get-AzDataProtectionImmutabilityRequestPipeline `
+                -ImmutabilityType $ImmutabilityType `
+                -ImmutabilityDurationInDay $ImmutabilityDurationInDay `
+                -HasImmutabilityType $hasImmutabilityType `
+                -HasImmutabilityDurationInDay $hasImmutabilityDurationInDay
+
+            if ($PSBoundParameters.ContainsKey("HttpPipelinePrepend")) {
+                $PSBoundParameters["HttpPipelinePrepend"] = $PSBoundParameters["HttpPipelinePrepend"] + @($immutabilityPipeline)
+            }
+            else {
+                $PSBoundParameters.Add("HttpPipelinePrepend", @($immutabilityPipeline))
+            }
+        }
+
         $hasToken = $PSBoundParameters.Remove("Token")
         $hasSecureToken = $PSBoundParameters.Remove("SecureToken")
         if($hasToken -or $hasSecureToken)
@@ -172,6 +208,7 @@ function Update-AzDataProtectionBackupVault
         $hasCmkUserAssignedIdentityId = $PSBoundParameters.Remove("CmkUserAssignedIdentityId")
         $hasCmkEncryptionKeyUri = $PSBoundParameters.Remove("CmkEncryptionKeyUri")
         $hasCostManagementGranularity = $PSBoundParameters.Remove("CostManagementGranularity")
+        if ($hasCostManagementGranularity) { $PSBoundParameters.Add("CostManagementSettingGranularityLevel", $CostManagementGranularity) }
 
         if (-not $hasCmkEncryptionState -and -not $hasCmkIdentityType -and -not $hasCmkUserAssignedIdentityId -and -not $hasCmkEncryptionKeyUri) {
             Az.DataProtection.Internal\Update-AzDataProtectionBackupVault @PSBoundParameters
@@ -214,7 +251,6 @@ function Update-AzDataProtectionBackupVault
         if ($hasEnableSystemAssignedIdentity) { $PSBoundParameters.Add("EnableSystemAssignedIdentity", $EnableSystemAssignedIdentity) }
         if ($hasAzureMonitorAlertsForAllJobFailure) { $PSBoundParameters.Add("AzureMonitorAlertsForAllJobFailure", $AzureMonitorAlertsForAllJobFailure) }
         if ($hasImmutabilityState) { $PSBoundParameters.Add("ImmutabilityState", $ImmutabilityState) }
-        if ($hasCostManagementGranularity) { $PSBoundParameters.Add("CostManagementSettingGranularityLevel", $CostManagementGranularity) }
         if ($hasCrossRegionRestoreState) { $PSBoundParameters.Add("CrossRegionRestoreState", $CrossRegionRestoreState) }
         if ($hasCrossSubscriptionRestoreState) { $PSBoundParameters.Add("CrossSubscriptionRestoreState", $CrossSubscriptionRestoreState) }
         if ($hasSoftDeleteRetentionDurationInDay) { $PSBoundParameters.Add("SoftDeleteRetentionDurationInDay", $SoftDeleteRetentionDurationInDay) }