+++ src/DataProtection/DataProtection.Autorest/custom/Helpers/CommonHelpers.ps1
@@ -133,3 +133,106 @@ function UnprotectSecureString
 		return $plaintext
 	}
 }
+
+function Assert-AzDataProtectionImmutabilitySetting
+{
+	[Microsoft.Azure.PowerShell.Cmdlets.DataProtection.DoNotExportAttribute()]
+	param(
+		[System.String]
+		${ImmutabilityState},
+
+		[System.String]
+		${ImmutabilityType},
+
+		[System.Boolean]
+		${HasImmutabilityState},
+
+		[System.Boolean]
+		${HasImmutabilityType},
+
+		[System.Boolean]
+		${HasImmutabilityDurationInDay}
+	)
+
+	if ($HasImmutabilityState -and $ImmutabilityState -eq 'Disabled' -and ($HasImmutabilityType -or $HasImmutabilityDurationInDay)) {
+		throw 'ImmutabilityType and ImmutabilityDurationInDay cannot be specified when ImmutabilityState is Disabled.'
+	}
+
+	if ($HasImmutabilityState -and $ImmutabilityState -ne 'Disabled' -and -not $HasImmutabilityType) {
+		throw 'ImmutabilityType is required when ImmutabilityState is Unlocked or Locked.'
+	}
+
+	if (($HasImmutabilityType -or $HasImmutabilityDurationInDay) -and -not $HasImmutabilityState) {
+		throw 'ImmutabilityState is required when ImmutabilityType or ImmutabilityDurationInDay is specified.'
+	}
+
+	if ($HasImmutabilityType -and $ImmutabilityType -eq 'TimeBased' -and -not $HasImmutabilityDurationInDay) {
+		throw 'ImmutabilityDurationInDay is required when ImmutabilityType is TimeBased.'
+	}
+
+	if ($HasImmutabilityType -and $ImmutabilityType -eq 'AsPerPolicy' -and $HasImmutabilityDurationInDay) {
+		throw 'ImmutabilityDurationInDay cannot be specified when ImmutabilityType is AsPerPolicy.'
+	}
+}
+
+function Get-AzDataProtectionImmutabilityRequestPipeline
+{
+	[Microsoft.Azure.PowerShell.Cmdlets.DataProtection.DoNotExportAttribute()]
+	param(
+		[System.String]
+		${ImmutabilityType},
+
+		[System.Nullable[System.Double]]
+		${ImmutabilityDurationInDay},
+
+		[System.Boolean]
+		${HasImmutabilityType},
+
+		[System.Boolean]
+		${HasImmutabilityDurationInDay}
+	)
+
+	$immutabilityTypeValue = $ImmutabilityType
+	$immutabilityDurationValue = $ImmutabilityDurationInDay
+	$includeImmutabilityType = $HasImmutabilityType
+	$includeImmutabilityDuration = $HasImmutabilityDurationInDay
+
+	$pipelineScript = {
+		param($request, $callback, $next)
+
+		if ($null -eq $request.Content) {
+			return $next.SendAsync($request, $callback)
+		}
+
+		$requestBody = $request.Content.ReadAsStringAsync().GetAwaiter().GetResult() | ConvertFrom-Json
+		if ($null -eq $requestBody.properties) {
+			$requestBody | Add-Member -MemberType NoteProperty -Name properties -Value ([PSCustomObject]@{})
+		}
+		if ($null -eq $requestBody.properties.securitySettings) {
+			$requestBody.properties | Add-Member -MemberType NoteProperty -Name securitySettings -Value ([PSCustomObject]@{})
+		}
+		if ($null -eq $requestBody.properties.securitySettings.immutabilitySettings) {
+			$requestBody.properties.securitySettings | Add-Member -MemberType NoteProperty -Name immutabilitySettings -Value ([PSCustomObject]@{})
+		}
+		if ($null -eq $requestBody.properties.securitySettings.immutabilitySettings.configuration) {
+			$requestBody.properties.securitySettings.immutabilitySettings |
+				Add-Member -MemberType NoteProperty -Name configuration -Value ([PSCustomObject]@{})
+		}
+
+		if ($includeImmutabilityType) {
+			$requestBody.properties.securitySettings.immutabilitySettings.configuration |
+				Add-Member -MemberType NoteProperty -Name type -Value $immutabilityTypeValue -Force
+		}
+		if ($includeImmutabilityDuration) {
+			$requestBody.properties.securitySettings.immutabilitySettings.configuration |
+				Add-Member -MemberType NoteProperty -Name durationInDays -Value $immutabilityDurationValue -Force
+		}
+
+		$jsonBody = $requestBody | ConvertTo-Json -Depth 100 -Compress
+		$request.Content = [System.Net.Http.StringContent]::new($jsonBody, [System.Text.Encoding]::UTF8, 'application/json')
+
+		return $next.SendAsync($request, $callback)
+	}.GetNewClosure()
+
+	return [Microsoft.Azure.PowerShell.Cmdlets.DataProtection.Runtime.SendAsyncStep]$pipelineScript
+}