+++ src/DataProtection/DataProtection.Autorest/test/utils.ps1
@@ -169,15 +169,15 @@ function setupEnv() {
 
     $AksRestoreVariables = @{
         SubscriptionId = "2c6832bf-90ef-457a-a663-9aeae88a7f80" 
-        ResourceGroupName = "tiering-test"
-        VaultName = "aks-tiering-test" #"demobackupvault"
-        NewPolicyName = "" #"pstest-aks-policy"
-        PolicyName = "test01"
+        ResourceGroupName = "AKSV-Vault-eastasia-LongHaul"
+        VaultName = "VaultProduction"
+        NewPolicyName = ""
+        PolicyName = "D5Retention"
         DataSourceLocation = "eastasia"
-        SourceClusterId = "/subscriptions/2c6832bf-90ef-457a-a663-9aeae88a7f80/resourceGroups/tiering-test/providers/Microsoft.ContainerService/managedClusters/aks-test-ps-backup"
-        SnapshotResourceGroupId = "/subscriptions/2c6832bf-90ef-457a-a663-9aeae88a7f80/resourceGroups/tiering-test"
-        FriendlyName = "aks-test-ps-backup\newBI"
-        ClusterName = "aks-test-ps-backup"
+        SourceClusterId = "/subscriptions/2c6832bf-90ef-457a-a663-9aeae88a7f80/resourceGroups/AKSV-Cluster-eastasia-LongHaul/providers/Microsoft.ContainerService/managedClusters/azk8s-eastasia-LongHaul"
+        SnapshotResourceGroupId = "/subscriptions/2c6832bf-90ef-457a-a663-9aeae88a7f80/resourceGroups/MC_azk8s-eastasia-LongHaul"
+        FriendlyName = "azk8s-eastasia-LongHaul"
+        ClusterName = "azk8s-eastasia-LongHaul"
     }
 
     $BlobHardeningVariables = @{
@@ -329,7 +329,7 @@ function setupEnv() {
     $DPPVaultSoftDeleteVariables = @{
         SubscriptionId = "38304e13-357e-405e-9e9a-220351dcce8c"
         Location = "eastus2euap" 
-        DeletedVaultName = "raghavchugh-test-dpp-uami-soft-delete-ON"
+        DeletedVaultName = "azps-dpp-vault-undelete-test"
         ResourceGroupName = "raghavchugh-test-rg"
     }
 
@@ -372,4 +372,3 @@ function setupEnv() {
 function cleanupEnv() {
     # Clean resources you create for testing
 }
-